V0.18

+++ api/product/garbagecolector.py
@@ -8,46 +8,47 @@ from django.db.models import Q
 class GarbageColectorProduct:
     def __init__(self):
         # Get an instance of a logger
-        self.logger = logging.getLogger()
+        self.logger = logging.getLogger("garbage_colector")
 
     def purge_products_expiration_time(self):
-
-        print("Remover todos os produtos que tem mais tempo do que o limit")
-
         try:
             expiration_time = settings.PRODUCT_EXPIRATION_TIME
-
-            print("Expiration Time: %s h" % expiration_time)
 
         except Exception as e:
             raise Exception("The PRODUCT_EXPIRATION_TIME variable is not configured in settings.")
 
-        today = datetime.today()
-        print("Today: %s" % today)
+        if expiration_time is not None:
 
-        date_expired = today - timedelta(hours=expiration_time)
-        print("Expired: %s" % date_expired)
+            self.logger.info("---------------------------------------------------------")
 
-        print("Todos os Produtos: %s" % Product.objects.all().count())
+            self.logger.info("Remove all products that have time greater than the limit")
+            self.logger.info("Expiration Time: %s h" % expiration_time)
 
-        # recuperar todos os produtos que foram criados a mais tempo que o expiration_time
-        # e que nao estejam marcados como permanente
-        products = Product.objects.filter(
-            Q(Q(prd_date__lt=date_expired) |
-              Q(prd_date__isnull=True)),
-            prd_is_permanent=False,
-            table__tbl_name__isnull=False
-        )
+            today = datetime.today()
+            date_expired = today - timedelta(hours=expiration_time)
 
-        print("Produtos a Serem removidos: %s" % products.count())
+            self.logger.info("Expired: %s" % date_expired)
 
-        for product in products:
-            print("Removendo produto: ID[%s] NAME[%s]" % (product.pk, product.prd_name))
-            try:
+            self.logger.info("Total Products: %s" % Product.objects.all().count())
 
-                product.delete()
-                print("Produto removido")
+            # recuperar todos os produtos que foram criados a mais tempo que o expiration_time
+            # e que nao estejam marcados como permanente
+            products = Product.objects.filter(
+                Q(Q(prd_date__lt=date_expired) |
+                  Q(prd_date__isnull=True)),
+                prd_is_permanent=False,
+                table__tbl_name__isnull=False
+            )
 
-            except Exception as e:
-                print(e)
-                print("falha ao remover produto")
+            self.logger.info("Products to be removed: %s" % products.count())
+
+            for product in products:
+                self.logger.info("Removing product: ID[%s] NAME[%s]" % (product.pk, product.prd_name))
+                try:
+
+                    product.delete()
+                    self.logger.info("Removed: ID[%s]" % product.pk)
+
+                except Exception as e:
+                    self.logger.warning(e)
+
